Fixed a naming convention error.
