+++ DataStructures/HashTable/HashTables.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.26124.0
 MinimumVisualStudioVersion = 15.0.26124.0
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HashTable", "HashTable\HashTable.csproj", "{31F97C88-1F8A-4760-B29A-0AEB30BE8065}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HashTables", "HashTables\HashTables.csproj", "{31F97C88-1F8A-4760-B29A-0AEB30BE8065}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
