+++ DataStructures/HashTable/HashTables/Program.cs
@@ -1,12 +1,19 @@
 ﻿using System;
+using HashTables.Classes;
 
-namespace HashTable
+namespace HashTables
 {
     class Program
     {
         static void Main(string[] args)
         {
             Console.WriteLine("Hello World!");
+
+            HashTable table = new HashTable(1024);
+
+            table.Add("soup", "tomato");
+
+            
         }
     }
 }
